requirements.txt
python-dotenv
dropbox
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
cryptography
pyTelegramBotAPI
pcloud
pyzipper
mega.py
tqdm
pyotp
telebot
pexpect
bitwarden-sdk
requests
argon2-cffi
caldav
icalendar
pyncclient
art
arrow
colorama
boto3
pytz
pykeepass
terminaltexteffects
tenacity==8.0.1
pycryptodome